# Optimized quoted-printable encoder